ok i must finish this for my wife's project

+++ CartService/CartService.BLL/Dtos/ProductDto.cs
@@ -12,8 +12,8 @@ namespace CartService.BLL.Dtos
         public string Name { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
         public decimal Price { get; set; }
-        public byte Rating { get; set; } = 0;
-        public long OrderAmount { get; set; } = 0;
+        public float Rating { get; set; } = 0;
+        public ulong OrderAmount { get; set; } = 0;
         public List<TagDto> Tags { get; set; } = new List<TagDto>();
         public string? ImageId { get; set; } = null;
     }
