+++ CartService/CartService/Dtos/Product/ProductResponseDto.cs
@@ -8,8 +8,8 @@ namespace CartService.Dtos.Product
         public string Name { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
         public decimal Price { get; set; }
-        public byte Rating { get; set; } = 0;
-        public long OrderAmount { get; set; } = 0;
+        public float Rating { get; set; } = 0;
+        public ulong OrderAmount { get; set; } = 0;
         public List<TagResponseDto> Tags { get; set; } = new List<TagResponseDto>();
         public string? ImageId { get; set; } = null;
     }
