+++ CatalogService/CatalogService.BLL/Dtos/ProductDto.cs
@@ -8,8 +8,8 @@ public class ProductDto
     public string Name { get; set; } = string.Empty;
     public string Description { get; set; } = string.Empty;
     public decimal Price { get; set; } = 0;
-    public byte Rating { get; set; } = 0;
-    public long OrderAmount { get; set; } = 0;
+    public float Rating { get; set; } = 0;
+    public ulong OrderAmount { get; set; } = 0;
     public List<TagDto> Tags { get; set; } = new List<TagDto>();
     public string? ImageId { get; set; } = null;
 }
